refactor: rename to telemetry
