+++ .env.example
@@ -1,6 +1,6 @@
 APP_MODE=production
-DATA_STREAM_SERVER_URL=http://127.0.0.1:8000/v1
-BLOCKERA_DATA_STREAM_OFF=false
+TELEMETRY_SERVER_URL=http://127.0.0.1:8000/v1/telemetry
+BLOCKERA_TELEMETRY_OFF=false
 WP_TESTS_DIR=/Users/User/Desktop/workflow/wordpress-develop/tests/phpunit/
 DB=wp_tests
