+++ .github/workflows/build-plugin-zip-tests.yml
@@ -48,7 +48,7 @@ jobs:
                   mkdir packages/blockera
                   mkdir packages/blockera/tests
                   touch .wp-env.json
-                  echo '{"plugins": ["."], "config": {"BLOCKERA_DATA_STREAM_OFF": true}}' >> .wp-env.json
+                  echo '{"plugins": ["."], "config": {"BLOCKERA_TELEMETRY_OFF": true}}' >> .wp-env.json
                   cat .wp-env.json
                   touch cypress.env.json
                   echo '{"isLogin": false,"wpUsername": "admin","wpPassword": "password","testURL": "http://localhost:8888","e2e": {"excludeSpecPattern": [],"specPattern": ["packages/**/*.build.e2e.cy.js"]}}' >> cypress.env.json
