+++ .wp-env.example.json
@@ -6,6 +6,6 @@
 		"https://downloads.wordpress.org/plugin/icon-block.latest-stable.zip"
 	],
 	"config": {
-		"BLOCKERA_DATA_STREAM_OFF": true
+		"BLOCKERA_TELEMETRY_OFF": true
 	}
 }
