+++ config/assets.php
@@ -39,13 +39,13 @@ return [
 			'data',
 			'icons',
 			'controls',
-			'data-stream',
+			'telemetry',
 			'bootstrap',
 			'wordpress',
 			'blockera-admin',
 			'controls-styles',
 			'wordpress-styles',
-			'data-stream-styles',
+			'telemetry-styles',
 			'blockera-admin-styles',
 		],
 		'with-deps' => [],
