+++ packages/blockera-admin/js/dashboard.js
@@ -30,7 +30,7 @@ import {
  */
 import { Panel } from './panel';
 import { config as optionsConfig } from './config';
-import { OptInModal } from '@blockera/data-stream';
+import { OptInModal } from '@blockera/telemetry';
 
 const getCurrentPage = (): string => {
 	const location = window.location;
@@ -59,8 +59,8 @@ export const Dashboard = (): MixedElement => {
 
 	return (
 		<div className={'blockera-settings-dashboard'}>
-			{'1' === window.blockeraDataStreamIsOff && (
-				<OptInModal kind={'blockera/v1'} name={'opt-in'} />
+			{'1' === window.blockeraTelemetryIsOff && (
+				<OptInModal kind={'blockera/v1'} name={'telemetry/opt-in'} />
 			)}
 			<SettingsContext.Provider
 				value={{
