+++ packages/blockera-admin/js/index.js
@@ -54,8 +54,8 @@ const initializeBlockeraAdmin = (): (() => void) => {
 				{
 					label: __('Blockera Opt-In', 'blockera'),
 					kind: 'blockera/v1',
-					name: 'opt-in',
-					baseURL: '/blockera/v1/opt-in',
+					name: 'telemetry/opt-in',
+					baseURL: '/blockera/v1/telemetry/opt-in',
 				},
 			]);
