--- packages/data-stream/composer.json
@@ -1,18 +1,0 @@
-{
-	"name": "blockera/data-stream",
-	"version": "0.9.2",
-	"description": "The data-stream system for blockera products by collecting necessary data of user and website to connect with Blockera API.",
-	"autoload": {
-		"psr-4": {
-			"Blockera\\DataStream\\": "php/"
-		},
-		"files": [
-			"php/functions.php"
-		]
-	},
-	"autoload-dev": {
-		"psr-4": {
-			"Blockera\\DataStream\\Test\\": "php/tests/"
-		}
-	}
-}
