+++ packages/telemetry/.npmrc
