+++ packages/telemetry/composer.json
@@ -1,0 +1,18 @@
+{
+	"name": "blockera/telemetry",
+	"version": "0.9.2",
+	"description": "The telemetry system for blockera products by collecting necessary data of user and website to connect with Blockera API.",
+	"autoload": {
+		"psr-4": {
+			"Blockera\\Telemetry\\": "php/"
+		},
+		"files": [
+			"php/functions.php"
+		]
+	},
+	"autoload-dev": {
+		"psr-4": {
+			"Blockera\\Telemetry\\Test\\": "php/tests/"
+		}
+	}
+}
