+++ packages/telemetry/js/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export * from './components/opt-in';
