+++ packages/telemetry/package.json
@@ -1,18 +1,18 @@
 {
-	"name": "@blockera/data-stream",
+	"name": "@blockera/telemetry",
 	"version": "0.9.2",
-	"description": "The data-stream system for blockera products to connect with api.blockera.ai site with beautiful UI.",
+	"description": "The telemetry system for blockera products to connect with api.blockera.ai site with beautiful UI.",
 	"author": "The  Contributors",
 	"license": "ISC",
 	"keywords": [
 		"blockera",
-		"data-stream"
+		"telemetry"
 	],
-	"homepage": "https://github.com/blockeraai/blockera/blob/master/packages/data-stream/README.md",
+	"homepage": "https://github.com/blockeraai/blockera/blob/master/packages/telemetry/README.md",
 	"repository": {
 		"type": "git",
 		"url": "https://github.com/blockeraai/blockera",
-		"directory": "packages/data-stream"
+		"directory": "packages/telemetry"
 	},
 	"bugs": {
 		"url": "https://github.com/blockeraai/blockera/issues"
