+++ packages/telemetry/php/DataProviders/DebugDataProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Blockera\DataStream\DataProviders;
+namespace Blockera\Telemetry\DataProviders;
 
 class DebugDataProvider implements DataProvider {
